analyzer/requirements.txt
# gRPC and Protocol Buffers
grpcio==1.66.2
grpcio-tools==1.66.2
protobuf==5.28.3

# Microsoft Presidio for PII Detection
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355

# NLP models for Presidio
spacy==3.7.6
https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl

# Additional ML/NLP dependencies
transformers==4.46.2
torch==2.5.1
sentence-transformers==3.3.1  # For lightweight jailbreak detection

# Async support
asyncio==3.4.3
aioredis==2.0.1

# Logging and monitoring
structlog==24.4.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-grpc==0.48b0
opentelemetry-exporter-prometheus==0.48b0

# Configuration
python-dotenv==1.0.1
pydantic==2.10.1
pydantic-settings==2.6.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-grpc==0.8.0
pytest-cov==6.0.0

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0
